requirements.txt
opencv-python==4.7.0.68
vidgear==0.3.0
argparse==1.1
matplotlib==3.6.3
numpy==1.23.4
matplotlib==3.6.2
numpy==1.23.5
openvino==2022.3.0
Pillow==9.5.0
torch==1.13.1
torchvision==0.14.1
vidgear==0.3.0
yt-dlp==2023.3.4
keras==2.11.0
openvino==2023.0.0
tensorflow==2.12.0
tensorflow_intel==2.11.0

